backend_brt-main/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pgvector==0.2.3
langchain==0.0.335
pandas==2.1.2
openpyxl==3.1.2
python-docx==1.0.1
